egs/atepp/local/reaper_requirements.txt
# Requirements for Reaper MIDI to Audio scripts

# reapy - Python wrapper for Reaper API (for reaper_midi_to_audio.py)
python-reapy>=0.9.0

# Optional but recommended
numpy>=1.20.0
